requirements.txt
# Ultra-Accurate Trading Bot Dependencies

# Core data science and machine learning
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Async and HTTP
asyncio-mqtt>=0.11.0
aiohttp>=3.8.0
aiofiles>=0.8.0

# Telegram bot
python-telegram-bot>=20.0

# Plotting and visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Technical analysis
TA-Lib>=0.4.24
talib-binary>=0.4.19  # Alternative for TA-Lib installation issues

# Deep learning (optional)
tensorflow>=2.8.0
torch>=1.11.0

# Utilities
python-dotenv>=0.19.0
colorama>=0.4.4
rich>=12.0.0
click>=8.0.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.18.0
black>=22.0.0
flake8>=4.0.0

# Financial data (optional)
yfinance>=0.1.70
alpha-vantage>=2.3.1

# Database (optional)
sqlite3  # Built-in
sqlalchemy>=1.4.0

# Logging and monitoring
loguru>=0.6.0

# Configuration
pyyaml>=6.0
toml>=0.10.2

# Performance
numba>=0.56.0
cython>=0.29.0